extract segment and block structs into spyre_allocator.h
